new file for generateUUID
